Move navigation service out of auth folder Fixes #61
