+++ test_driver/mocks/services/mock_nav_service.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:mockito/mockito.dart';
-import 'package:niira/services/auth/navigation_service.dart';
+import 'package:niira/services/navigation_service.dart';
 
 class MockNavService extends Mock implements NavigationService {
   @override
